Add support for showing "footsteps" on the map
Adds a config which allows the user to specify the length (in tiles) of
a trail to be rendered on the world map showing where some GIMP has
walked. The color of the trail depends on the index of the GIMP in his
group (the colors for now are hardcoded).

Also changes the smooth map point motion logic to smoothly move even
between different planes. I just realized that "planes" seem to mostly
refer to different stories/levels within the same area. Moving to a
completely different area (e.g. ZMI) doesn't seem to affect the plane.